+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpConstants.java
@@ -436,6 +436,16 @@ public class HttpConstants {
     public static final BString SECURESOCKET_CONFIG_CERT_VALIDATION_TYPE_OCSP_STAPLING =
             StringUtils.fromString("OCSP_STAPLING");
 
+    //Socket Config
+    public static final BString SOCKET_CONFIG = StringUtils.fromString("socketConfig");
+    public static final BString SOCKET_CONFIG_SO_BACKLOG = StringUtils.fromString("soBackLog");
+    public static final BString SOCKET_CONFIG_CONNECT_TIMEOUT = StringUtils.fromString("connectTimeOut");
+    public static final BString SOCKET_CONFIG_RECEIVE_BUFFER_SIZE = StringUtils.fromString("receiveBufferSize");
+    public static final BString SOCKET_CONFIG_SEND_BUFFER_SIZE = StringUtils.fromString("sendBufferSize");
+    public static final BString SOCKET_CONFIG_TCP_NO_DELAY = StringUtils.fromString("tcpNoDelay");
+    public static final BString SOCKET_CONFIG_SOCKET_REUSE = StringUtils.fromString("socketReuse");
+    public static final BString SOCKET_CONFIG_KEEP_ALIVE = StringUtils.fromString("keepAlive");
+
     //Client Endpoint (CallerActions)
     public static final String CLIENT_ENDPOINT_SERVICE_URI = "url";
     public static final String CLIENT_ENDPOINT_CONFIG = "config";