+++ native/src/main/java/io/ballerina/stdlib/http/transport/contractimpl/DefaultHttpWsConnectorFactory.java
@@ -144,9 +144,9 @@ public class DefaultHttpWsConnectorFactory implements HttpWsConnectorFactory {
     }
 
     @Override
-    public HttpClientConnector createHttpClientConnector(
-            Map<String, Object> transportProperties, SenderConfiguration senderConfiguration) {
-        BootstrapConfiguration bootstrapConfig = new BootstrapConfiguration(transportProperties);
+    public HttpClientConnector createHttpClientConnector(Map<String, Object> transportProperties,
+                                                         SenderConfiguration senderConfiguration) {
+        BootstrapConfiguration bootstrapConfig = new BootstrapConfiguration(senderConfiguration);
         ConnectionManager connectionManager = new ConnectionManager(senderConfiguration.getPoolConfiguration());
         int configHashCode = Util.getIntProperty(transportProperties, HttpConstants.CLIENT_CONFIG_HASH_CODE, 0);
         return new DefaultHttpClientConnector(connectionManager, senderConfiguration, bootstrapConfig, clientGroup,
@@ -154,10 +154,10 @@ public class DefaultHttpWsConnectorFactory implements HttpWsConnectorFactory {
     }
 
     @Override
-    public HttpClientConnector createHttpClientConnector(
-        Map<String, Object> transportProperties, SenderConfiguration senderConfiguration,
-        ConnectionManager connectionManager) {
-        BootstrapConfiguration bootstrapConfig = new BootstrapConfiguration(transportProperties);
+    public HttpClientConnector createHttpClientConnector(Map<String, Object> transportProperties,
+                                                         SenderConfiguration senderConfiguration,
+                                                         ConnectionManager connectionManager) {
+        BootstrapConfiguration bootstrapConfig = new BootstrapConfiguration(senderConfiguration);
         int configHashCode = Util.getIntProperty(transportProperties, HttpConstants.CLIENT_CONFIG_HASH_CODE, 0);
         return new DefaultHttpClientConnector(connectionManager, senderConfiguration, bootstrapConfig, clientGroup,
                                               configHashCode);