+++ native/src/test/java/io/ballerina/stdlib/http/transport/contract/config/ServerBootstrapConfigurationTest.java
@@ -43,23 +43,21 @@ public class ServerBootstrapConfigurationTest {
     @Test
     public void testIsSocketReuse() {
         ServerBootstrapConfiguration serverBootstrapConfiguration = new ServerBootstrapConfiguration(new HashMap<>());
-        Assert.assertFalse(serverBootstrapConfiguration.isSocketReuse());
+        Assert.assertTrue(serverBootstrapConfiguration.isSocketReuse());
 
         HashMap<String, Object> properties = new HashMap<>();
-        properties.put(Constants.SERVER_BOOTSTRAP_SO_REUSE, true);
+        properties.put(Constants.SERVER_BOOTSTRAP_SO_REUSE, false);
         serverBootstrapConfiguration = new ServerBootstrapConfiguration(properties);
-        Assert.assertTrue(serverBootstrapConfiguration.isSocketReuse());
+        Assert.assertFalse(serverBootstrapConfiguration.isSocketReuse());
     }
 
     @Test
     public void testGetSoTimeout() {
         ServerBootstrapConfiguration serverBootstrapConfiguration = new ServerBootstrapConfiguration(new HashMap<>());
-        Assert.assertEquals(serverBootstrapConfiguration.getSoTimeOut(), 15);
-
+        Assert.assertEquals(serverBootstrapConfiguration.getConnectTimeOut(), 15000);
         HashMap<String, Object> properties = new HashMap<>();
-        properties.put(Constants.SERVER_BOOTSTRAP_SO_TIMEOUT, 10);
+        properties.put(Constants.SERVER_BOOTSTRAP_CONNECT_TIME_OUT, 10000);
         serverBootstrapConfiguration = new ServerBootstrapConfiguration(properties);
-        Assert.assertEquals(serverBootstrapConfiguration.getSoTimeOut(), 10);
+        Assert.assertEquals(serverBootstrapConfiguration.getConnectTimeOut(), 10000);
     }
-
 }